Updated __str__ representation of an optimization.
Formatting changed so that output is in more consistent columns.

Type of constraints now correctly displays 'e' for equality constraints
or 'i' for inequality constraints.

Added an index to each variable and constraint to make it easier to
decipher optimizer information about variables and constraints.

Added a status column to variables and constraints that gives information
regarding the bounds.  This makes it more obvious when a solution is
being limited by a bound, and when convergence fails it makes it more obvious
which constraints are not being met.  The status shows one or more of the
following:

l - Lower bound active
u - Upper bound active
L - Lower bound violated
U - Upper bound violated
E - Equality constraint violated

Currently the threshold for being active or in violation is 1.0E-6.  Ultimately
this should probably be dependent on the tolerance settings of the optimizer.