Restored lagrange multipliers in the string representation of the solution.
Added lagrange multipliers to IPOPT's output. Note that in the call to _createSolution we negate the multiplier values so that they are consistent with those returned by SNOPT.
My understanding is that this is a matter of convention (due to the internal representations of the problems in the two optimizers).